requirements.txt
Flask==2.0.2
waitress==2.0.0
gunicorn==20.1.0
tensorflow-cpu==2.7.0
numpy==1.19.5
scipy==1.7.3
mediapipe==0.8.9